requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
alabaster==0.7.12         # via sphinx
amqp==2.5.2               # via kombu
appdirs==1.4.3            # via black, virtualenv
appnope==0.1.0            # via ipython
argon2-cffi==19.2.0       # via -r requirements.in
asgiref==3.2.7            # via django
astroid==2.4.0            # via pylint, pylint-celery
attrs==19.3.0             # via black, pytest
babel==2.8.0              # via sphinx
backcall==0.1.0           # via ipython
billiard==3.6.3.0         # via celery
black==19.10b0            # via -r requirements.in
boto3==1.12.47            # via django-storages
botocore==1.15.47         # via boto3, s3transfer
celery==4.4.2             # via -r requirements.in, django-celery-beat, flower
certifi==2020.4.5.1       # via requests, sentry-sdk
cffi==1.14.0              # via argon2-cffi
cfgv==3.1.0               # via pre-commit
chardet==3.0.4            # via requests
click==7.1.2              # via black, uvicorn
collectfast==2.1.0        # via -r requirements.in
coverage==5.1             # via -r requirements.in, django-coverage-plugin
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via python3-openid
distlib==0.3.0            # via virtualenv
django-allauth==0.41.0    # via -r requirements.in
django-anymail[mailgun]==7.1.0  # via -r requirements.in
django-celery-beat==2.0.0  # via -r requirements.in
django-coverage-plugin==1.8.0  # via -r requirements.in
django-crispy-forms==1.9.0  # via -r requirements.in
django-debug-toolbar==2.2  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-extensions==2.2.9  # via -r requirements.in
django-filter==2.2.0      # via -r requirements.in
django-model-utils==4.0.0  # via -r requirements.in
django-redis==4.11.0      # via -r requirements.in
django-storages[boto3]==1.9.1  # via -r requirements.in, collectfast
django-stubs==1.5.0       # via -r requirements.in
django-timezone-field==4.0  # via django-celery-beat
django==3.0.5             # via -r requirements.in, collectfast, django-allauth, django-anymail, django-celery-beat, django-debug-toolbar, django-filter, django-model-utils, django-redis, django-storages, django-stubs, django-timezone-field, djangorestframework, model-bakery
djangorestframework==3.11.0  # via -r requirements.in
docutils==0.15.2          # via botocore, sphinx
entrypoints==0.3          # via flake8
factory-boy==2.12.0       # via -r requirements.in
faker==4.0.3              # via factory-boy
filelock==3.0.12          # via virtualenv
flake8-isort==3.0.0       # via -r requirements.in
flake8==3.7.9             # via -r requirements.in, flake8-isort
flower==0.9.4             # via -r requirements.in
gunicorn==20.0.4          # via -r requirements.in
h11==0.9.0                # via uvicorn
httptools==0.1.1          # via uvicorn
humanize==0.5.1           # via flower
identify==1.4.15          # via pre-commit
idna==2.9                 # via requests
imagesize==1.2.0          # via sphinx
ipdb==0.13.2              # via -r requirements.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.13.0           # via ipdb
isort[pyproject]==4.3.21  # via flake8-isort, pylint
jedi==0.17.0              # via ipython
jinja2==2.11.2            # via sphinx
jmespath==0.9.5           # via boto3, botocore
kombu==4.6.8              # via celery
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8, pylint
model-bakery==1.1.0       # via -r requirements.in
more-itertools==8.2.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.770               # via -r requirements.in, django-stubs
nodeenv==1.3.5            # via pre-commit
oauthlib==3.1.0           # via requests-oauthlib
packaging==20.3           # via pytest, pytest-sugar, sphinx
parso==0.7.0              # via jedi
pathspec==0.8.0           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.1.2             # via -r requirements.in
pluggy==0.13.1            # via pytest
pre-commit==2.3.0         # via -r requirements.in
prompt-toolkit==3.0.5     # via ipython
psycopg2==2.8.5           # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pycparser==2.20           # via cffi
pyflakes==2.1.1           # via flake8
pygments==2.6.1           # via ipython, sphinx
pylint-celery==0.3        # via -r requirements.in
pylint-django==2.0.15     # via -r requirements.in
pylint-plugin-utils==0.6  # via pylint-celery, pylint-django
pylint==2.5.0             # via pylint-celery, pylint-django, pylint-plugin-utils
pyparsing==2.4.7          # via packaging
pytest-django==3.9.0      # via -r requirements.in
pytest-sugar==0.9.3       # via -r requirements.in
pytest==5.4.1             # via -r requirements.in, pytest-django, pytest-sugar
python-crontab==2.4.2     # via django-celery-beat
python-dateutil==2.8.1    # via botocore, faker, python-crontab
python-slugify==4.0.0     # via -r requirements.in
python3-openid==3.1.0     # via django-allauth
pytz==2020.1              # via -r requirements.in, babel, celery, django, django-timezone-field, flower
pyyaml==5.3.1             # via pre-commit
redis==3.4.1              # via -r requirements.in, django-redis
regex==2020.4.4           # via black
requests-oauthlib==1.3.0  # via django-allauth
requests==2.23.0          # via django-allauth, django-anymail, requests-oauthlib, sphinx
s3transfer==0.3.3         # via boto3
sentry-sdk==0.14.3        # via -r requirements.in
six==1.14.0               # via argon2-cffi, astroid, django-anymail, django-coverage-plugin, django-extensions, packaging, python-dateutil, traitlets, virtualenv
snowballstemmer==2.0.0    # via sphinx
sphinx==3.0.3             # via -r requirements.in
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
sqlparse==0.3.1           # via django, django-debug-toolbar
termcolor==1.1.0          # via pytest-sugar
testfixtures==6.14.1      # via flake8-isort
text-unidecode==1.3       # via faker, python-slugify
toml==0.10.0              # via black, isort, pre-commit, pylint
tornado==6.0.4            # via flower
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.2  # via collectfast, django-stubs, mypy
urllib3==1.25.9           # via botocore, requests, sentry-sdk
uvicorn==0.11.3           # via -r requirements.in
uvloop==0.14.0            # via uvicorn
vine==1.3.0               # via amqp, celery
virtualenv==20.0.18       # via pre-commit
wcwidth==0.1.9            # via prompt-toolkit, pytest
websockets==8.1           # via uvicorn
werkzeug==1.0.1           # via -r requirements.in
wrapt==1.12.1             # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
